% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_of_composition.R
\name{beta_bayes}
\alias{beta_bayes}
\title{Apply fully Bayesian version of method of composition}
\usage{
beta_bayes(stmobj, formula, metadata, nsims = 100, seed = 123)
}
\arguments{
\item{stmobj}{Fitted stm model.}

\item{formula}{Formula (subset of the prevalence specification used to fit the stm).}

\item{metadata}{Metadata that was used to fit the stm.}

\item{nsims}{Number of repetitions.}

\item{seed}{Seed.}
}
\value{
A list of lists: For each topic, nsims regression outputs are returned.
}
\description{
For a fitted stm, first sample topic proportions from the approximate prosterior,
and then perform a Bayesian beta regression of these proportions on prevalence covariates.
The procedure is repeated nsims times.
}
