% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_of_composition.R
\name{posterior_predict_props1}
\alias{posterior_predict_props1}
\title{Generate mean and credible intervals of topic proportions over full range of a specified variable}
\usage{
posterior_predict_props1(
  bayes_out,
  est_var,
  formula,
  metadata,
  ci_lower,
  ci_upper
)
}
\arguments{
\item{bayes_out}{Output of method of composition with Bayesian regression using beta_bayes}

\item{est_var}{Variable for which to sample over full observed range.}

\item{formula}{Formula (subset of the prevalence specification used to fit the stm).}

\item{metadata}{Metadata that was used to fit the stm.}

\item{ci_lower}{Lower bound of credible interval.}

\item{ci_upper}{Upper bound of credible interval.}
}
\value{
A list of dataframes: For each topic, the empirical mean and credible intervals are returned.
}
\description{
Given the MAP estimates obtained using beta_bayes, specify one variable and predict
mean and credible intervals of topic proportions over the full observed range of variable est_var,
while holding all other variables as median/majority. This function is typically used to visualize
the results of the method of composition.
}
