% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_of_composition.R
\name{predict_props}
\alias{predict_props}
\title{Generate mean and credible intervals of topic proportions over full range of a specified variable}
\usage{
predict_props(
  beta_coefs,
  est_var,
  formula,
  metadata,
  ci_lower = 0.025,
  ci_upper = 0.975
)
}
\arguments{
\item{beta_coefs}{sampled coefficients from method of composition using sample_coefs}

\item{est_var}{Variable for which to sample over full observed range.}

\item{formula}{Formula (subset of the prevalence specification used to fit the stm).}

\item{metadata}{Metadata that was used to fit the stm.}

\item{ci_lower}{Lower bound of credible interval.}

\item{ci_upper}{Upper bound of credible interval.}
}
\value{
A list of dataframes: For each topic, the empirical mean and credible intervals are returned.
}
\description{
Given the samples of coefficients obtained using sample_coefs, specify one variable and predict
mean and credible intervals (with respect to mean prediction) of topic proportions
over the full observed range of variable est_var,
while holding all other variables as median/majority. This function is typically used to visualize
the results of the method of composition.
}
