% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_of_composition.R
\name{sample_coefs}
\alias{sample_coefs}
\title{Perform Method of Composition employing either a quasibinomial GLM or beta regression}
\usage{
sample_coefs(stmobj, formula, type, metadata, nsims = 25, seed = NULL)
}
\arguments{
\item{stmobj}{Fitted stm model.}

\item{formula}{Formula (subset of the prevalence specification used to fit the stm).}

\item{type}{Regression to perform: Either 'beta' or 'quasibinomial'.}

\item{metadata}{Metadata that was used to fit the stm.}

\item{nsims}{Number of repetitions.}

\item{seed}{Seed.}
}
\value{
A list of lists: For each topic, nsims regression outputs are returned.
}
\description{
For a fitted stm, first sample topic proportions from the approximate prosterior,
then perform a regression of these proportions on prevalence covariates,
and lastly sample from the resulting distribution of regression coefficents.
The obtained values of this process, which is repeated nsims times,
are samples of the marginal posterior of regression coefficients.
This procedure is known as the method of composition in the social sciences.
}
