% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_props_logisticn.R
\name{sample_props_logisticn}
\alias{sample_props_logisticn}
\title{Sample from LogisticNormal over full range of variable}
\usage{
sample_props_logisticn(
  stmobj,
  est_var,
  formula,
  metadata,
  nsims = 1000,
  ci_lower = 0.025,
  ci_upper = 0.975,
  seed = NULL
)
}
\arguments{
\item{stmobj}{Fitted stm model.}

\item{est_var}{Variable for which to sample over full observed range.}

\item{formula}{Formula (the prevalence specification used to fit the stm).}

\item{metadata}{Metadata that was used to fit the stm.}

\item{nsims}{Number of draws from the LogisticNormal used to obtain empirical mean and quantiles.}

\item{ci_lower}{Lower bound of credible interval.}

\item{ci_upper}{Upper bound of credible interval.}

\item{seed}{Seed.}
}
\value{
A list of dataframes: For each topic, the empirical mean and credible intervals are returned.
}
\description{
Sample mean and credible interval from LogisticNormal for a given model (stmobj)
and full observed range of variable (est_var),
holding all variables but est_var as median/majority.
}
